+++ gems/aws-sdk-globalaccelerator/lib/aws-sdk-globalaccelerator.rb
@@ -48,6 +48,6 @@ require_relative 'aws-sdk-globalaccelerator/customizations'
 # @!group service
 module Aws::GlobalAccelerator
 
-  GEM_VERSION = '1.40.0'
+  GEM_VERSION = '1.41.0'
 
 end