+++ gems/aws-sdk-resiliencehub/lib/aws-sdk-resiliencehub/client.rb
@@ -1560,7 +1560,7 @@ module Aws::ResilienceHub
     #   resp.component_recommendations[0].config_recommendations[0].description #=> String
     #   resp.component_recommendations[0].config_recommendations[0].ha_architecture #=> String, one of "MultiSite", "WarmStandby", "PilotLight", "BackupAndRestore", "NoRecoveryPlan"
     #   resp.component_recommendations[0].config_recommendations[0].name #=> String
-    #   resp.component_recommendations[0].config_recommendations[0].optimization_type #=> String, one of "LeastCost", "LeastChange", "BestAZRecovery", "LeastErrors", "BestAttainable"
+    #   resp.component_recommendations[0].config_recommendations[0].optimization_type #=> String, one of "LeastCost", "LeastChange", "BestAZRecovery", "LeastErrors", "BestAttainable", "BestRegionRecovery"
     #   resp.component_recommendations[0].config_recommendations[0].recommendation_compliance #=> Hash
     #   resp.component_recommendations[0].config_recommendations[0].recommendation_compliance["DisruptionType"].expected_compliance_status #=> String, one of "PolicyBreached", "PolicyMet"
     #   resp.component_recommendations[0].config_recommendations[0].recommendation_compliance["DisruptionType"].expected_rpo_description #=> String
@@ -1830,6 +1830,7 @@ module Aws::ResilienceHub
     #   resp.app_summaries[0].description #=> String
     #   resp.app_summaries[0].name #=> String
     #   resp.app_summaries[0].resiliency_score #=> Float
+    #   resp.app_summaries[0].status #=> String, one of "Active", "Deleting"
     #   resp.next_token #=> String
     #
     # @see http://docs.aws.amazon.com/goto/WebAPI/resiliencehub-2020-04-30/ListApps AWS API Documentation
@@ -2705,11 +2706,6 @@ module Aws::ResilienceHub
     #   The type of resiliency policy to be created, including the recovery
     #   time objective (RTO) and recovery point objective (RPO) in seconds.
     #
-    #   <note markdown="1"> If you do not want to specify regional targets for a regional policy,
-    #   you must set the values of `rpoInSecs` and `rtoInSecs` to `-1`.
-    #
-    #    </note>
-    #
     # @option params [required, String] :policy_arn
     #   The Amazon Resource Name (ARN) of the resiliency policy. The format
     #   for this ARN is:
@@ -2788,7 +2784,7 @@ module Aws::ResilienceHub
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-resiliencehub'
-      context[:gem_version] = '1.6.0'
+      context[:gem_version] = '1.7.0'
       Seahorse::Client::Request.new(handlers, context)
     end