+++ gems/aws-sdk-cloudwatchrum/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.5.0 (2022-10-20)
+------------------
+
+* Feature - CloudWatch RUM now supports Extended CloudWatch Metrics with Additional Dimensions
+
 1.4.0 (2022-02-24)
 ------------------